big fixed

+++ zipmo.ps1
@@ -13,7 +13,7 @@ if ([string]::IsNullOrEmpty($output)) {
 
 if ([string]::IsNullOrEmpty($fiducial)) {
     if (Test-Path '..\last_zipped_time.txt') {
-        $fiducial = Get-Content last_zipped_time.txt
+        $fiducial = Get-Content '..\last_zipped_time.txt'
         $fiducial = $fiducial.Trim()
     } else {
         Write-Output 'Specify fiducial date and time like "2023-04-06T16:20:00".'
@@ -26,7 +26,9 @@ if (Test-Path $fileName) {
 }
 $modifiedFiles = @(Get-ChildItem $filePattern -File | Where-Object {$_.LastWriteTime -gt $fiducial} | Select-Object -ExpandProperty Name)
 zip.exe $fileName $modifiedFiles
-Move-Item -Force $fileName ..
+if (Test-Path $fileName) {
+    Move-Item -Force $fileName ..
+}
 
 $lastZippedTime = get-date -format "yyyy-MM-ddTHH:mm:ss"
 Set-Content -Path '..\last_zipped_time.txt' $lastZippedTime
